refactor: group Flame map icon caches under caches submodule (#3878)
Move AssetImageCache and marker icon caches into features/game/flame/caches/
with a public barrel export, update imports and repo-lint paths, and keep
import-boundary and decode-dedup gates green.

Refs #3878
