+++ SPEC/program/repo-lint.md
@@ -51,7 +51,7 @@
 | `tool/check_debug_console_logic_contract_boundary.dart` | AST-enforced scoped package import contract for debug-console → logic boundary; rule `repo.debug_console_logic_contract_boundary` |
 | `tool/check_debug_console_shared_helpers.dart` | Positive invariants for shared debug-console parse/execute helpers (`SPEC/program/debug-console-internals.md`); rule `repo.debug_console_shared_helpers` |
 | `tool/check_app_debug_handler_guard_helpers.dart` | Forbid inline duplicated guard string literals in `app/lib/core/services/app_event_handler_debug_*.dart`; handlers must use the shared guard helpers in `debug_command_helpers.dart` (`SPEC/program/debug-console-internals.md`); rule `repo.app_debug_handler_guard_helpers` |
-| `tool/check_app_asset_image_decode_dedup.dart` | AST scan over `app/lib/**` enforcing the shared Flame asset-image decode helper (Refs #3699): the `ui.decodeImageFromList(...)` PNG-decode idiom may appear **only** in the canonical helper file `app/lib/features/game/flame/asset_image_cache.dart` (which exposes `decodeImageAsset(...)`); every icon cache and tileset must route decode through that helper instead of re-inlining `rootBundle.load` -> `Completer` -> `decodeImageFromList`. Additionally, every concrete icon cache (a public class declared in a `*_icon_cache.dart` file whose name ends with `IconCache`) must declare `extends AssetImageCache` so the load lifecycle / `_isLoading`/`_isLoaded` flags / structured logging / decode fan-out stay single-sourced in the base; the `AssetImageCache` base file is exempt and tilesets adopt only the decode helper (not the cache base). Generated suffixes, `_test.dart` / `/test/` paths, and fixture dirs are skipped; matches use AST method-invocation / extends-clause checks so comments are ignored; rule `repo.app_asset_image_decode_dedup` |
+| `tool/check_app_asset_image_decode_dedup.dart` | AST scan over `app/lib/**` enforcing the shared Flame asset-image decode helper (Refs #3699): the `ui.decodeImageFromList(...)` PNG-decode idiom may appear **only** in the canonical helper file `app/lib/features/game/flame/caches/asset_image_cache.dart` (which exposes `decodeImageAsset(...)`); every icon cache and tileset must route decode through that helper instead of re-inlining `rootBundle.load` -> `Completer` -> `decodeImageFromList`. Additionally, every concrete icon cache (a public class declared in a `*_icon_cache.dart` file whose name ends with `IconCache`) must declare `extends AssetImageCache` so the load lifecycle / `_isLoading`/`_isLoaded` flags / structured logging / decode fan-out stay single-sourced in the base; the `AssetImageCache` base file is exempt and tilesets adopt only the decode helper (not the cache base). Generated suffixes, `_test.dart` / `/test/` paths, and fixture dirs are skipped; matches use AST method-invocation / extends-clause checks so comments are ignored; rule `repo.app_asset_image_decode_dedup` |
 | `tool/check_app_event_handler_scope_logic_boundary.dart` | Enforce that `app/lib/core/services/app_event_handler_scope.dart` has no direct import from `app/lib/features/game/logic/**`; rule `repo.app_event_handler_scope_logic_boundary` |
 | `tool/check_subscription_tracker.dart` | Enforce that `app/lib/features/**/*.dart` does not store subscriptions in raw `StreamSubscription` lists; use `SubscriptionTracker` for shared-lifecycle subscription cleanup; rule `repo.subscription_tracker` |
 | `tool/check_app_lib_no_suggest_work_orders.sh` | Forbid `suggestWorkOrders(` call sites under `app/lib/**` (per-unit UI uses `getAvailableWorkTargetsForUnit` / `getValidWorkOrderTileKeysWithVisibility`; Refs #2133); rule `repo.app_lib_no_broad_suggest_work_orders` |
@@ -369,7 +369,7 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given chrome button sources under `app/lib/features/game/widgets/chrome/**`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.chrome_button_base`, then every `State<X>` whose `X` is a public class name ending in `TextButton` declares `with CtHoverButtonStateMixin`, and the run fails listing each violating `path:line` when such a state omits the canonical mixin (Refs #3594).
 - Given in-game split-dialog sources under `app/lib/features/game/widgets/**`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.split_dialog_base`, then every public class whose name starts with `Split` and ends with `Dialog` (other than the `SplitEntityDialog` base) declares `extends SplitEntityDialog`, and the run fails listing each violating `path:line` when such a dialog omits the canonical base (Refs #3594).
 - Given in-game train-at-capital dialog sources under `app/lib/features/game/widgets/**`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.train_dialog_base`, then every public class whose name starts with `Train` and ends with `Dialog` (other than the `TrainDialogBase` base) declares `extends TrainDialogBase`, and the run fails listing each violating `path:line` when such a dialog omits the canonical base (Refs #3594).
-- Given Flame asset-image sources under `app/lib/**`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_asset_image_decode_dedup`, then `ui.decodeImageFromList(...)` is invoked only in `app/lib/features/game/flame/asset_image_cache.dart` and every public `*IconCache` class declared in a `*_icon_cache.dart` file declares `extends AssetImageCache`, and the run fails listing each violating `path:line` when a file outside the helper re-inlines the decode idiom or an icon cache omits the canonical base (Refs #3699).
+- Given Flame asset-image sources under `app/lib/**`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_asset_image_decode_dedup`, then `ui.decodeImageFromList(...)` is invoked only in `app/lib/features/game/flame/caches/asset_image_cache.dart` and every public `*IconCache` class declared in a `*_icon_cache.dart` file declares `extends AssetImageCache`, and the run fails listing each violating `path:line` when a file outside the helper re-inlines the decode idiom or an icon cache omits the canonical base (Refs #3699).
 - Given the commodity-cost train dialog sources `app/lib/features/game/widgets/train_military_dialog.dart` and `train_naval_dialog.dart`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_train_dialog_commodity_cost_shared`, then each file declares a state class `extends CommodityCostTrainDialogState` and re-declares none of `_totalTreasuryCost` / `_totalPeasantCost` / `_totalCommodityCosts` / `_remainingTreasury` / `_remainingPeasants` / `_remainingCommodity` / `_deficitHint` / `canAffordIncrement` nor any private `_*ResourceBar` / `_*Row` widget, and the run fails listing each violating `path:line` when a dialog omits the canonical base or re-duplicates the cost machinery (Refs #3686).
 - Given repository Dart files excluding generated suffixes (`.g.dart`, `.freezed.dart`, `.mocks.dart`, `.gen.dart`) and generated app l10n paths under `app/lib/l10n/gen/` matching the checker’s prefix list, when `dart run tool/ct_repo_lint.dart` runs rule `repo.dart_file_non_comment_line_size`, then each remaining scanned file has at most 1000 non-comment lines and the run fails while listing every violating file when any file is strictly greater than 1000.
 - Given the [Policy: no violation allowlists](#policy-no-violation-allowlists-repo-lint) section, when `repo.repeated_magic_numbers` runs, then that checker does not load keyed waiver data; a decoy keyed-waiver-shaped YAML file under `tool/` must not suppress fail-level repeated-literal violations for in-scope domain sources (see `SPEC/program/repeated-magic-numbers.md` and `test/check_repeated_magic_numbers_test.dart`).
