+++ app/lib/features/game/flame/caches/caches.dart
@@ -1,0 +1,9 @@
+/// Flame map marker and tileset image caches (Refs #3878 Phase 3).
+library;
+
+export 'asset_image_cache.dart';
+export 'civilian_icon_cache.dart';
+export 'fleet_icon_cache.dart';
+export 'province_label_icon_cache.dart';
+export 'resource_icon_cache.dart';
+export 'town_icon_cache.dart';
