+++ app/lib/features/game/flame/caches/province_label_icon_cache.dart
@@ -1,6 +1,6 @@
 import 'dart:ui' as ui;
 
-import '../../../config/app_assets.dart';
+import '../../../../config/app_assets.dart';
 import 'asset_image_cache.dart';
 
 const Set<String> kProvinceLabelIconIds = {
