+++ app/lib/features/game/flame/caches/town_icon_cache.dart
@@ -1,6 +1,6 @@
 import 'dart:ui' as ui;
 
-import '../../../config/app_assets.dart';
+import '../../../../config/app_assets.dart';
 import 'asset_image_cache.dart';
 
 /// All town icon cache ids: 16 level/style variants plus the port glyph.
