+++ app/lib/features/game/flame/transport_overlay_tileset.dart
@@ -5,7 +5,7 @@ import 'package:colonizethis_app/config/map_terrain_config.dart';
 import 'package:colonizethis_app/package_logger.dart';
 import 'package:flutter/services.dart';
 
-import 'asset_image_cache.dart';
+import 'caches/asset_image_cache.dart';
 
 final _log = packageLogger();
