+++ test/check_app_asset_image_decode_dedup_test.dart
@@ -140,6 +140,6 @@ void main() {
 }
 
 Directory _flameDir(Directory temp) {
-  return Directory(p.join(temp.path, 'app/lib/features/game/flame'))
+  return Directory(p.join(temp.path, 'app/lib/features/game/flame/caches'))
     ..createSync(recursive: true);
 }
